+++ json_serializable/test/src/_json_serializable_test_input.dart
@@ -27,6 +27,8 @@ part 'setter_test_input.dart';
 
 part 'to_from_json_test_input.dart';
 
+part 'unknown_enum_value_test_input.dart';
+
 part 'unknown_type_test_input.dart';
 
 @ShouldThrow('Generator cannot target `theAnswer`.',